gubookhub_project/requirements.txt
certifi==2020.12.5
chardet==4.0.0
Django==2.2.17
django-registration-redux==2.2
django-widget-tweaks==1.4.8
idna==2.10
Pillow==8.1.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
